refactor: Split data type definitions into separate files.
