+++ src/data-types/bitn.js
@@ -1,0 +1,6 @@
+module.exports = {
+  id: 0x68,
+  type: 'BITN',
+  name: 'BitN',
+  dataLengthLength: 1
+};
