+++ src/data-types/datetimen.js
@@ -1,0 +1,6 @@
+module.exports = {
+  id: 0x6F,
+  type: 'DATETIMN',
+  name: 'DateTimeN',
+  dataLengthLength: 1
+};
