+++ src/data-types/decimaln.js
@@ -1,0 +1,8 @@
+module.exports = {
+  id: 0x6A,
+  type: 'DECIMALN',
+  name: 'DecimalN',
+  dataLengthLength: 1,
+  hasPrecision: true,
+  hasScale: true
+};
