+++ src/data-types/floatn.js
@@ -1,0 +1,6 @@
+module.exports = {
+  id: 0x6D,
+  type: 'FLTN',
+  name: 'FloatN',
+  dataLengthLength: 1
+};
