+++ src/data-types/intn.js
@@ -1,0 +1,6 @@
+module.exports = {
+  id: 0x26,
+  type: 'INTN',
+  name: 'IntN',
+  dataLengthLength: 1
+};
