+++ src/data-types/money.js
@@ -1,0 +1,36 @@
+const MoneyN = require('./moneyn');
+
+module.exports = {
+  id: 0x3C,
+  type: 'MONEY',
+  name: 'Money',
+
+  declaration: function() {
+    return 'money';
+  },
+
+  writeTypeInfo: function(buffer) {
+    buffer.writeUInt8(MoneyN.id);
+    buffer.writeUInt8(8);
+  },
+
+  writeParameterData: function(buffer, parameter) {
+    if (parameter.value != null) {
+      buffer.writeUInt8(8);
+      buffer.writeMoney(parameter.value * 10000);
+    } else {
+      buffer.writeUInt8(0);
+    }
+  },
+
+  validate: function(value) {
+    if (value == null) {
+      return null;
+    }
+    value = parseFloat(value);
+    if (isNaN(value)) {
+      return new TypeError('Invalid number.');
+    }
+    return value;
+  }
+};
