+++ src/data-types/moneyn.js
@@ -1,0 +1,6 @@
+module.exports = {
+  id: 0x6E,
+  type: 'MONEYN',
+  name: 'MoneyN',
+  dataLengthLength: 1
+};
