+++ src/data-types/ntext.js
@@ -1,0 +1,9 @@
+module.exports = {
+  id: 0x63,
+  type: 'NTEXT',
+  name: 'NText',
+  hasCollation: true,
+  hasTableName: true,
+  hasTextPointerAndTimestamp: true,
+  dataLengthLength: 4
+};
