+++ src/data-types/null.js
@@ -1,0 +1,5 @@
+module.exports = {
+  id: 0x1F,
+  type: 'NULL',
+  name: 'Null'
+};
