+++ src/data-types/numericn.js
@@ -1,0 +1,8 @@
+module.exports = {
+  id: 0x6C,
+  type: 'NUMERICN',
+  name: 'NumericN',
+  dataLengthLength: 1,
+  hasPrecision: true,
+  hasScale: true
+};
