+++ src/data-types/real.js
@@ -1,0 +1,36 @@
+const FloatN = require('./floatn');
+
+module.exports = {
+  id: 0x3B,
+  type: 'FLT4',
+  name: 'Real',
+
+  declaration: function() {
+    return 'real';
+  },
+
+  writeTypeInfo: function(buffer) {
+    buffer.writeUInt8(FloatN.id);
+    buffer.writeUInt8(4);
+  },
+
+  writeParameterData: function(buffer, parameter) {
+    if (parameter.value != null) {
+      buffer.writeUInt8(4);
+      buffer.writeFloatLE(parseFloat(parameter.value));
+    } else {
+      buffer.writeUInt8(0);
+    }
+  },
+
+  validate: function(value) {
+    if (value == null) {
+      return null;
+    }
+    value = parseFloat(value);
+    if (isNaN(value)) {
+      return new TypeError('Invalid number.');
+    }
+    return value;
+  }
+};
