+++ src/data-types/udt.js
@@ -1,0 +1,6 @@
+module.exports = {
+  id: 0xF0,
+  type: 'UDTTYPE',
+  name: 'UDT',
+  hasUDTInfo: true
+};
