+++ src/data-types/xml.js
@@ -1,0 +1,6 @@
+module.exports = {
+  id: 0xF1,
+  type: 'XML',
+  name: 'Xml',
+  hasSchemaPresent: true
+};
